# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/me": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/tipos-documento-fisico": {
      "get": {
        "operationId": "TiposDocumentoFisicoController_listar",
        "parameters": [
          {
            "name": "activo",
            "required": false,
            "in": "query",
            "description": "Filtrar por estado activo. Omitir para listar solo activos.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Búsqueda parcial sobre nombre.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListarTiposDocumentoFisicoResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lista los tipos de documento físico del tenant activo. Orden: tributarios primero, luego por nombre.",
        "tags": [
          "Tipos de Documento Físico"
        ]
      },
      "post": {
        "operationId": "TiposDocumentoFisicoController_crear",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTipoDocumentoFisicoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Crea un tipo de documento físico en el tenant activo.",
        "tags": [
          "Tipos de Documento Físico"
        ]
      }
    },
    "/api/tipos-documento-fisico/{id}": {
      "patch": {
        "operationId": "TiposDocumentoFisicoController_actualizar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTipoDocumentoFisicoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "PATCH del tipo. Campos opcionales; solo toca los que vienen. El campo `codigo` es inmutable y se ignora si se envía.",
        "tags": [
          "Tipos de Documento Físico"
        ]
      },
      "delete": {
        "operationId": "TiposDocumentoFisicoController_eliminar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Elimina un tipo de documento físico. Falla con 409 si tiene documentos físicos asociados (usar desactivar en ese caso).",
        "tags": [
          "Tipos de Documento Físico"
        ]
      }
    },
...
  },
  "info": {
    "title": "Multi-Tenant SaaS API",
    "description": "Production-ready NestJS API with multi-tenancy, authentication, RBAC, billing, and more.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      },
      "X-Tenant-ID": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-ID",
        "description": "Tenant ID for multi-tenant operations"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "example": "John Doe",
            "description": "Display name"
          }
        }
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "isEmailVerified",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "TipoDocumentoFisicoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "codigo": {
            "type": "string"
          },
          "esTributario": {
            "type": "boolean"
          },
          "activo": {
            "type": "boolean"
          },
          "tiposComprobanteAplicables": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "APERTURA",
                "DIARIO",
                "INGRESO",
                "EGRESO",
                "AJUSTE",
                "TRASPASO",
                "CIERRE"
              ]
            }
          },
          "organizationId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre",
          "codigo",
          "esTributario",
          "activo",
          "tiposComprobanteAplicables",
          "organizationId",
          "createdAt",
          "updatedAt"
        ]
      },
      "ListarTiposDocumentoFisicoResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TipoDocumentoFisicoResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "CreateTipoDocumentoFisicoDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "example": "Factura recibida",
            "minLength": 1,
            "maxLength": 100
          },
          "codigo": {
            "type": "string",
            "example": "factura-recibida",
            "maxLength": 20,
            "description": "Identificador estable en kebab-case alfanumérico. Inmutable post-creación.",
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
          },
          "esTributario": {
            "type": "boolean",
            "example": false,
            "description": "true para documentos con requisitos tributarios (factura, NC, ND)."
          },
          "tiposComprobanteAplicables": {
            "type": "array",
            "example": [
              "EGRESO",
              "DIARIO"
            ],
            "description": "Tipos de comprobante con los que este tipo puede asociarse. Array vacío = ninguno.",
            "items": {
              "type": "string",
              "enum": [
                "APERTURA",
                "DIARIO",
                "INGRESO",
                "EGRESO",
                "AJUSTE",
                "TRASPASO",
                "CIERRE"
              ]
            }
          }
        },
        "required": [
          "nombre",
          "codigo",
          "esTributario",
          "tiposComprobanteAplicables"
        ]
      },
      "UpdateTipoDocumentoFisicoDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "esTributario": {
            "type": "boolean",
            "description": "true para documentos tributarios (factura, NC, ND)."
          },
          "activo": {
            "type": "boolean",
            "description": "Activar o desactivar el tipo sin eliminarlo."
          },
          "tiposComprobanteAplicables": {
            "type": "array",
            "description": "Reemplaza la lista completa de tipos de comprobante aplicables.",
            "items": {
              "type": "string",
              "enum": [
                "APERTURA",
                "DIARIO",
                "INGRESO",
                "EGRESO",
                "AJUSTE",
                "TRASPASO",
                "CIERRE"
              ]
            }
          }
        }
      },
...
    }
  }
}