# backend/src/tipos-documento-fisico/dto/tipo-documento-fisico-response.dto.ts
import type { TipoDocumentoFisico } from '@prisma/client';
import { ApiProperty } from '@nestjs/swagger';
import { TipoComprobante } from '@prisma/client';

export class TipoDocumentoFisicoResponseDto {
  @ApiProperty() id!: string;
  @ApiProperty() nombre!: string;
  @ApiProperty() codigo!: string;
  @ApiProperty() esTributario!: boolean;
  @ApiProperty() activo!: boolean;
  @ApiProperty({ isArray: true, enum: TipoComprobante })
  tiposComprobanteAplicables!: TipoComprobante[];
  @ApiProperty() organizationId!: string;
  @ApiProperty() createdAt!: string;
  @ApiProperty() updatedAt!: string;
}

export class ListarTiposDocumentoFisicoResponseDto {
  @ApiProperty({ type: () => [TipoDocumentoFisicoResponseDto] })
  items!: TipoDocumentoFisicoResponseDto[];
  @ApiProperty() total!: number;
  @ApiProperty() page!: number;
  @ApiProperty() pageSize!: number;
}

export function toTipoDocumentoFisicoResponse(
  t: TipoDocumentoFisico,
): TipoDocumentoFisicoResponseDto {
  return {
    id: t.id,
    nombre: t.nombre,
    codigo: t.codigo,
    esTributario: t.esTributario,
    activo: t.activo,
    tiposComprobanteAplicables: t.tiposComprobanteAplicables,
    organizationId: t.organizationId,
    createdAt: t.createdAt.toISOString(),
    updatedAt: t.updatedAt.toISOString(),
  };
}